refactor: authStores로 파일 이름 변경
