+++ lib/store/useAuthStore.ts
@@ -1,6 +1,6 @@
 import { create } from 'zustand';
 import { createJSONStorage, persist } from 'zustand/middleware';
-import { AuthStore } from '../types/authStore';
+import { AuthStore } from '../types/authStores';
 
 const useAuthStore = create<AuthStore>()(
   persist(
